Address review feedback for PR #87
- Move CHANGELOG entry from released [1.9.1] to new [1.9.2] - Unreleased section
- Document the @SchemaProp::example drop as a Changed note
- Defensively skip non-object property values in Schema::propertySchemas()
  to avoid TypeError on JSON Schema boolean schemas when asserts are off
- Split JsonSchemaInputFixture.php into one-class-per-file (JsonSchemaInput
  extracted to its own file)
- Refresh docs/schemas/index.html and ship new json-schema-input.param.json
