+++ docs/schemas/index.html
@@ -16,6 +16,7 @@
 <li><a href="card.json" rel="schema">card.json</a></li>
 <li><a href="contact.param.json" rel="schema">contact.param.json</a></li>
 <li><a href="invalid.json" rel="schema">invalid.json</a></li>
+<li><a href="json-schema-input.param.json" rel="schema">json-schema-input.param.json</a></li>
 <li><a href="numbers.json" rel="schema">numbers.json</a></li>
 <li><a href="org.json" rel="schema">org.json</a></li>
 <li><a href="person.json" rel="schema">person.json</a></li>
